Delay-load Roslyn services
Commit bcb2c8c7 added an `[Import]` of `IDiagnosticAnalyzerService` to
our Solution Explorer code. Unfortunately, this causes us to compose and
load effectively all of Roslyn whenever the Solution Explorer is opened,
even if no C# or VB code is loaded.

The fix is to delay pulling in the `IDiagnosticAnalyzerService` until we
actually need to make use of it, which is guaranteed to be after we have
loaded a C# or VB project (and thus we've already loaded Roslyn).