Introduce a hoisted local scope abstraction
Since hoisted locals are fields, they are in scope throughout the entire
method body.  However, from the user's perspective, they are still locals,
so we attempt to retain their original scopes during debugging.
Unfortunately, C# and VB use different mechanisms for doing so.  In order
to make the code more consistent, and to simplify consumption of the new
PDB format (which is more similar to C#), we introduce an abstract type
InScopeHoistedLocals that represents the set of in-scope hoisted locals in
a consistent way in both languages (and both PDB formats).