Enable bootstrap builds
This adds a new supported MSBuild property BootstrapBuild.  The value of
this must point to a directory containing a functional version of csc
and vbc.  When specified these executables will be used to build the
product instead of the versions provided in NuGet.

This provides us a very simple way of enabling bootstrap builds.