Fix Linux build warnings
The SignVsixInputs target depends on the CopyFilesToOutputDirectory target.  This target logically exists on xbuild but is named DeployOutputFiles.  This descrepancy lead to numerous warnings on Linux during a compilation about missing targets.

After discussing with @jasonmalinowski we decided to move it to
AfterCompile.  This executes in the right place for SignVsixInputs and
is supported in both xbuild and msbuild.