Handle pseudo-variables with pointer types
The intrinsic methods that we call to retrieve the values of
pseudo-variables from the debugger return System.Object.  Unfortunately,
object cannot be converted (directly) to a pointer type.  Instead, we need
to unbox an IntPtr and then call an explicit conversion method to get a
pointer.

Caveat: This won't work until the debugger actually returns a boxed
IntPtr.