Make test usage of ISymUnmanagedReader uniform
We used to have multiple mechanisms - constructing a co-class directly
from a registered GUID, wrapping the co-class in SymReader (for some extra
constant mocking), wrapping the co-class in a TempPdbReader (IDisposable),
and wrapping a TempPdbReader in a DebugInfoProvider (trivial wrapper).

This change eliminates TempPdbReader and DebugInfoProvider and gives
SymReader exclusive access to the raw co-class.

This will be useful when we switch to ISymUnmanagedReader3, which the
co-class does not presently implement (but which SymReader can).