Change the build client heuristic for using a running server process
This implementation changes the compiler server named pipe to use
the SHA1 hash of the compiler exe directory as the named pipe name.
The client then attempts to connect to the pipe with this name and,
if it is unable, starts the server, which creates and connects to
a pipe of said name. If all else fails, continue with fallback
compilation as usual.