Only access VS services when required
There are a couple of MEF-instantiated types that immediately obtain VS
services in their constructors. This can cause us to load and run code
from other assemblies that we may never actually use, and certainly
don't need right away.

Instead, we should just hold on to the `IServiceProvider` and use it to
obtain the VS service as-needed.