Add support for "dynamic" in the result provider
This is the first of a sequence of changes.  We associate additional
information with each type that comes from the expression compiler.  For
now, it is the bool[] from the DynamicAttribute on the return type of a
method synthesized by the expression compiler.  (The bool[] is packed into
a byte[] for generality and compactness.)  Then, as we traverse the type
in the type name formatter, we consume the flags to determine which
occurrences of "object" should be replaced by "dynamic".

We don't show "dynamic" everywhere we could (e.g. in base type names in
disambiguators), but I believe we show "dynamic" everywhere the legacy EE
did.

TODO: The expression compiler doesn't call the new API to pass the
appropriate dynamic flags (API pending).

TODO: The expression compiler doesn't consume dynamic flags on
pseudo-locals (API pending).