Add a fallback for when the build task client is GAC'd.
Normally, the build client for the compiler server looks for the server
in the same directory as itself (the running executable), however when
the client is inside the build task, the task may be GAC'd and thus the
server may be nowhere near by.

If so, we should fallback to the location MSBuild found the client exe
in (usually the default MSBuild directory), and otherwise give up if
we aren't given a fallback path.