Add a Roslyn-wide rule set file.
While we have several different .ruleset files covering specific parts
of the Roslyn code (e.g., compiler projects, or projects that contain
analyzers, etc.) we do not currently have a rule set file that covers
all of Roslyn. This would be useful as a single place to turn off a
badly-behaved diagnostic, or turn on a new one that is off by default.

This commit adds such a file, and updates the other rule set files to
refer back to it. It also updates one of our .targets files to set this
as the default rule set automatically; this way we don't need to go back
and update every project to refer to it.

Fixes #1061.