Add the /analyzerdependency switch to the compilers
Currently we require that an analyzer's dependencies be deployed
immediately next to it in the file system, and the various analyzer
hosts (csc.exe, vbc.ex, VS, etc.) handle finding and loading these
dependencies automatically.

It has become clear, however, that this deployment requirement needs to
be relaxed. For example, some of our analyzer NuGets have C# analyzers
in a C# subdirectory, and VB analyzers in a VB subdirectory. If the C#
and VB analyzers both depend on another assembly, we now need to have an
identical copy in each subdirectory, which is silly.

Instead we should allow dependencies to be placed anywhere they make
sense, but to enable this we will need to know which .dlls are
dependencies and which are not. Hence the /analyzerdependency switch.

This commit just adds the switch to the compilers' command line parsing
so they don't fail when they see it. Later changes will pipe support
through MSBuild, the language services, and the VB and C# project system
support. Once all of that is in place the compilers will be updated to
only look for dependencies in the locations specified by the the switch.