Explicitly pass down compiler exe location
The common parts of the compiler were attempting to deduce the directory
of the compiler exe by using various reflection / assembly APIs.  This
is a fragile method that is easy to break in unrelated code.
Additionally this API is not available on CoreFx.

The fix is to simply thread the compiler exe path through from the Main
methods on the EXE.

closes #1705