Fix variable capturing for VB lambdas in EE...
Variables may have been captured by lambdas in the original method
or in the expression, and we need to preserve the existing values of
those variables in the expression. This requires rewriting the variables
in the expression based on the closure classes from both the original
method and the expression, and generating a preamble that copies
values into the expression closure classes.

Consider the original method:
```
Shared Sub M()
    Dim x, y, z as Integer
    ...
    F(Function() x + y)
End Sub
```
and the expression in the EE: "F(Function() x + z)".
The expression is first rewritten using the closure class and the local
`<1>` from the original method: F(Function() `<1>`.x + z)
Then lambda rewriting introduces a new closure class that includes
the locals `<1>` and z, and a corresponding local
`<2>`: F(Function() `<2>`.`<1>`.x + `<2>`.z)
And a preamble is added to initialize the fields of `<2>`:
```
    <2> = New <>c__DisplayClass0()
    <2>.<1> = <1>
    <2>.z = z
```