Fix an issue that was causing the rename tracking fix to offer up empty preview.
In cases where primary document from which the rename tracking fix was triggered doesn't have any changes (aside from the rename which is already in the buffer), we would end up offering no preview. This change ensures that we will go on and return preview for some other document that did have changes in such cases.