Fix NullReferenceException during debugging
When opening a file during a debugging session, VS sets its buffer read-only when it is not editable. To do so, VS asks its document to the workspace but in some cases, the workspace returns nothing, causing NRE and VS Crash. Issue #1202 is tracking to fix this issue directly (by enforcing the workspace to load projects before EnC). In this PR, we defensively add a null check and log it to make a further investigation easier.