Set correct path to linked rule set file
When the user right-clicks on a rule set file in Solution Explorer and
chooses "Set as Active Rule Set", we ask the associated `IVsHierarchy`
for the name of the rule set file and use that to set the
`<CodeAnalysisRuleSet>` property in the project file. However, the name
returned by the `IVsHierarchy` does not include the full path, so this
only works if the rule set is immediately next to the project file.

For linked files this will generally not be the case. Instead, we need
to ask for the *canonical* name, which will give us the full path. We
then compute the relative path to that file from the project's
directory, and use that for the `<CodeAnalysisRuleSet>` property.

Fixes #1579.