Specially handle root namespace constituents in SourceNamespaceSymbol.IsDefinedInSourceTree.
Adjust implementation of SourceNamespaceSymbol.IsDefinedInSourceTree to account for the fact that constituents of a root namespace are implicitly defined in every syntax tree and, therefore, should be visited for the purpose of collecting diagnostics within a syntax tree. Without this, an unexpected “unnecessary imports” diagnostic could be reported for a syntax tree and some expected diagnostics could remain unreported for the tree.