Pass the name of the pipe as the first argument to the compiler server.
Currently, a bug is present in the perf machine where the compiler
server parent directory and the client parent directory differ only by
casing. Rather than try to fix this specific instance, we should remove
the reliance on .NET APIs to return the same value for the client and the
server by simply passing the pipe name to the compiler server.

This should mean that every identical instantiation of the client results
in the same pipe name and remove the logic in the server.