Avoid a hang in CompileWithRetry...
If we've already asked the debugger to load an assembly with a particular identity, bail out if we ever try to load the same assembly again.  We expect that compilation should succeed (or generate a different error) after the first attempt to load the missing assembly.  If it doesn't, then something else is wrong, but we don't want to loop infinitely.