Explicitly pass SDK path to compilers
This changes the compiler APIs to no longer use
RuntimeVersion.GetRuntimeDirectory to find the path to mscorlib and
System.dll for the following reasons:

- This API is not available in CoreFx and hence blocks our transition.
- This behavior of implicitly referencing mscorlib doesn't have a
logical equivalent in coreclr where there may be no reachable mscorlib
on the machine (think Linux).

The compiler now takes the SDK search directory as a paramater.  The
consmer will be responsible for feeding down the correct value to the
compiler.

This is the first step in resolving issue #1939