Change how we do UTF8 output
Both compilers support the /utf8output flag to change the output
encoding to UTF8.  The implementation was using Console.OutputEncoding
which is not supported in CoreFx.  Changed the implementation to use the
supported API pairs of OpenStandard* and Set*.

Note: The compiler was also using IsOutputRedirected which isn't
supported on CoreFx and has no portable replacement.  After discussion
with the compiler team we decided to omit this check.  It a check that
is more aimed at preventing the user from making a mistake by not
redirecting the output vs. enforcing an actual constraint.