Print process output to console instead of results file
In certain error cases, like if xunit crashes, it may not produce a test results file.  Instead of writing the process output to the test results xml file, which could be improperly interpreted as xml, write it to the console.  Also, remove add Process.Start call in case of failed xunit process.