Add a basic assembly loader.
Currently we expect that an analyzer's dependencies are located next to
it in the file system, and we do not require that those dependencies be
explicitly passed to the compiler. These requirements are changing. We
will now allow analyzers and their dependencies to be located anywhere,
and this requires that the full set of analyzers and dependencies be
passed in so that we know where to find them.

In order to load analyzers and dependencies under these new
requirements, this commit introduces the SimpleAnalyzerAssemblyLoader.
This type loads assemblies with a simple `Assembly.LoadFrom`, but it
also hooks the `AppDomain.AssemblyResolve` event in order to find and
load items from its list of dependencies.

Csc.exe and vbc.exe have been updated to use this loader instead of
calling directly into `Assembly.LoadFrom`. It is also exposed to the
workspace layer through the IAnalyzerService type.