Port CA1027 and CA2217 (Mark\Do not mark Enums with FlagsAttribute) to System.Runtime.Analyzers
This is a straightforward move as the fixer was already built on SyntaxGenerator. It had C#\VB specific fixers though because for VB, to replace the node, it needed to know the distinction between block\statement. DocumentEditor.ReplaceNode abstracts that and so we can have a unified fixer. The analyzer was branching out for getting the location and that's not really needed.

Conflicts:
	src/Diagnostics/FxCop/System.Runtime.Analyzers/Core/SystemRuntimeAnalyzersResources.Designer.cs
	src/Diagnostics/FxCop/System.Runtime.Analyzers/Core/SystemRuntimeAnalyzersResources.resx