Remove a dependency on InMemoryAssemblyLoader
Previously, the `ServerDispatcher.AnalyzerWatcher` class would hook the
`InMemoryAssemblyLoader.AssemblyLoad` event. When it saw that a loaded
assembly's file had changed on disk it would signal the process stop
accepting new connections and shutdown when the current compilations
were complete. This minimized the window of time where the compiler
server might execute loaded-but-outdated analyzers during a compilation.

The `InMemoryAssemblyLoader` is going away, and we shouldn't depend on
being told of every assembly load anyway. Instead, `AnalyzerWatcher` has
become `CompilerServerFileWatcher`, and the `CSharpCompilerServer` and
`VisualBasicCompilerServer` are responsible for telling it which files
need to be watched based on the analyzer references passed in.