Remove a dependency on InMemoryAssemblyProvider
The `AnalyzerFileWatcherService` currently depends on the
`InMemoryAssemblyProvider` to tell it when it when an analyzer assembly
has been loaded so that it can watch for changes to the file on disk.
Instead, we can get this information directly from the language service.

This change also removes the code that tries to map from a dependency
back to some analyzer that directly or indirectly loaded it. In general
we can't rely on being able to create a chain back to an analyzer, and
it shouldn't be as much of a problem now that we require dependencies to
be declared.

Also, when adding or removing analyzer dependencies we need to kick off
the dependency analysis just like we do with analyzers.