Use different mechanism for passing symbol reader to MTA...
Previously, we'd store a direct pointer to the IUnknown instance and
unwrap that in the MTA.  The IENCSymbolReaderProvider interface is
being changed to prevent calls from an STA by returning the IUnknown
instance directly, rather than a pointer to it.

Thus, we need to add a layer of indirection (marshal the IUnknown
in a Stream and pass a pointer to the Stream instance to the MTA).