Self host vbc builds on Mac CI
This updates our cibuild.sh script to include self hosting the VBC
build in addition to the CSC build.

The support is limited to our Mac runs only for the moment due to Linux
not being upgrade to the proper mono runtime.   This will be updated in
the near future.