Add support for "dynamic" in the expression compiler
This is the second of a sequence of changes.  (The first was 3dd1f2b.)  In
this change we start returning custom type info (effectively just the
flags of DynamicAttribute packed into a byte array) from calls to the
expression compiler so that they can be consumed by the result provider.

TODO: The API for locals is still in flux and we don't presently have a
place to return the custom type info.

TODO: This change does not associate custom type info with locals declared
in the Immediate window.