Add support for "dynamic" Immediate window locals
This is the third of a sequence of changes (see also 3dd1f2b and f2cac22).
In this change we start consuming the custom type info (effectively just
the flags of DynamicAttribute packed into a byte array) associated with
locals declared in the Immediate window.

TODO: The GetAliases API has not been implemented, but this functionality
will light up when it is.