Change named pipe ownership check
The goal is to have a single csc.exe / vbc.exe binary that runs cross
plat and cross runtime (mono, desktop, coreclr).  Yet there are certain
features which only light up in a subset of the scenarios.  For example
the compiler server only lights up on Windows platforms.

This is fine but forces us to jump through hoops when doing platform
specific actions like checking the owner of the other end of the named
pipe.  To compile this code into a portable assembly we either need to
a) pinvoke or b) use reflection.  In this case reflection was the
simpler route.