Change Signature: Update all documents from the base solution
Fixes #784 and internal TFS issue 1102830

In change signature, when we needed to update 10 documents, we would
update the first document and replace it in the solution, then update
the second document and replace it in the solution, etc. However, if the
definition document is updated before a reference document, then trying
to bind the reference to the definition will not work correctly. Most of
the update process is syntactic so it usually doesn't matter, but we do
bind method references to see the invocation uses a reduced extension
method. This caused reduced extension method references to sometimes get
updated incorrectly.

We now produce all of the updated syntax trees from the same base
solution, and then apply them all at the end.