Increment the compiler client pipe name if one already exists
There are certain cases, e.g. when a user elevates their permission
level, when a user would like to use two concurrent compiler servers
started from the same path, but with different processes. Since the
current connection semantics depend on the pipe name being the hash
of the directory containing the client exe, this is not possible.

This change appends an incrementing integer to the requested pipe
name until an available pipe is found, and changes the connection
semantics for the client to search all pipe name variants until an
empty one is found or it has connected.