Display a tooltip for hyperlinks in error list and preview pane
Display a tooltip containing url destination for hyperlinks in the error list and light bulb preview pane UIs.

Also use full fidelity (i.e. substituted)  error message in the search query in the case of hyperlinks that redirect to bing search so that we can present more relevant search results.