When classifying xml nodes in VB, use FullSpan
Since the editor calls us line by line, if we get called to classify a
comment trivia on the line before an xml node, we would not do so as part
of that line, because it doesn't overlap the Span.  We would later do so,
but not actually return the spans to the editor, since they didn't
overlap the span requested.

Fixes #2126.