Only reopen documents on navigation when required
Fixes internal TFS bug #1063320 "Diff View: Navigating using NavBar in diff view opens new file"

When the VisualStudioDocumentNavigationService was updated to correctly handle NavigateTo's document tab state transition from Provisional to Permanent, the navigation process was updated to always reopen the document so that such a state transition could take place (and the new document state setting could be observed). However, performing this reopen on *all* navigation was causes other problems, such as opening a new window when the Navigation Bar is used in a diff view. We now limit this document re-opening behavior to scenarios that require it, which is currently only NavigateTo.