Use IEditorFormattingService to format on commit
Previously, if the user committed completion with semicolon (or another
auto-format triggering character), the completion commit code would run,
followed by the autoformat code. Both features run in places where
multiple edits happen sequentially, so they use
AutomaticCodeChangeMergePolicy when creating undo transactions. However,
this meant that commit followed by format-on-typechar would only produce
one undo transaction.

To avoid this, when committing, completion queries the relevant
IEditorFormattingService to determine if the commit character should
result in formatting, and then uses IEditorFormattingService to format
within a separate undo transaction with the default merge policy (the
command handler for format-on-typechar does essentially the same thing).

This results in two items in the undo stack: one for committing
completion, and one for the format. When the normal formatting command
handler runs, the formatter doesn't produce any changes, so it does
nothing.