Check analyzer consistency
The change adds the `ConsistencyChecker` type and wires it into csc.exe
and vbc.exe. This type validates that a set of analyzer assemblies on
disk and their loaded equivalents are complete and consistent.

This type performs two checks:

1. It checks that the assembly file on disk and the corresponding loaded
   assembly are the same. This is done by comparing the MVIDs values.
2. It checks that all of an assemblies dependencies are satisified by
   other assemblies in the set. E.g., if assembly A depends on assembly B,
   B must also be specified. A whitelist can be provided to skip checks for
   assemblies that are expected to be provided and loaded by the host, for
   example, mscorlib, System.*, and Microsoft.CodeAnalysis.*.

Together these rules catch cases where an analyzer may not have all the
dependencies it needs to run correctly, or where it is going to run with
different dependencies than expected.