Make constructor for Location internal
Fixes #1587

Location is currently a public abstract class. However, since Location is something that can be provided to a Roslyn Diagnostic (via Diagnostic.Create()) supporting arbitrary derived Location types means that Roslyn's analyzer and fixer engines would need to be resilient against arbitrary implemntations that can throw exceptions.

This change makes it impossible to inherit Location externally so that we don't have to do complicated work to guard against such implementations.

There are no known external inheritors of Location at the moment (as far as we are aware). Plus we always have the option to ease this restriction and allow inheritance in the future if we find that there are legitimate use cases for it.