Fix loading the type of FixedBufferAttribute in fields
The current code attempts to load an arbitrary type from the
FixedBufferAttribute, which represents the type of pointer which
is returned by the fixed buffer field in an unsafe struct. This is
a problem because it is not done with assembly identity unification
and the unification may depend on the current compilation, even though
the type is in metadata.

However, the type of the fixed buffer field is limited to SpecialTypes,
so rather than trying to do unification we can simply load the special
type of the field directly.

Fixes #2385