Replace d4a181 fix for FixedBufferAttribute with proper fix
The old fix used SpecialType decoding to fetch the unified special type
for the encoded type of a FixedBufferAttribute. While this works, the
underlying issue behind the missing type inside the FixedBufferAttribute
is that decoded types are loaded by checking the referenced assembly symbols
rather than the referenced assemblies.

This reverts the change from d4a181 and changes type decoding in
attribute loading to use the referenced assemblies.