Push `AnalyzerFileReference` into CodeAnalysis
Push the `AnalyzerFileReference` type out of the CodeAnalysis.Desktop
project and into the CodeAnalysis (Portable) project.

This is the primary type that we use to load diagnostic analyzers from
assemblies. It contains the logic to find analyzers within an assembly
and instantiate only those needed for a particular language (e.g., C# or
VB) and report any errors in doing so. As such, it makes sense to have
this type as widely available as possible. If it were left in the
CodeAnalysis.Desktop project this functionality would conceivably have
to be implemented by multiple different hosts.

We do lose the ability to validate the path to the assembly. However, we
don't actually use this path to load the assembly--that has to be
handled by the host. Instead, we use it for informational purposes, and
also to look for types within the assembly using the metadata reader. A
future commit will remove the latter use--at this point we always have
the assembly loaded before this type is instantiated, so there is no
point in using the metadata reader to avoid the cost of a load. We can
simply reflect over the assembly instead.