Re-enable insertion of System.Composition DLLs
We used to insert System.Composition.Convention.dll S.C.Hosting.dll, and
S.C.TypedParts.dll (even though they were in a list of assemblies to
specifically include).  When we stopped recently, we discovered that VS
was not providing them.  Resume inserting them until we figure out who is
supposed to be providing them.