Add an Id property to AnalyzerReference
Fixes #1710.

We can easily run into the situation where the same analyzer assembly is
appears twice: once as an `<AnalyzerReference>` in the project, and once
in a VSIX installed in VS. We don't want to run the analyzer twice, so
currently the `HostAnalyzerManager` in VS checks
`AnalyzerReference.DisplayName` and excludes any analyzers from VSIXes
that match analyzers from the project.

This works for the moment, but it isn't what `DisplayName` was designed
for. Instead, this commit adds an `Id` property to `AnalyzerReference`
and updates the `HostAnalyzerManager` to make use of it.