log watson when codefix or refactoring throws
When a codefix or refactoring throws an exception, we turn it off, log
it to the activity log, and put up and info bar telling the user what
happened.

With this change we also check to see if the codefix/refactoring was
implemented by the roslyn team and send a non-fatal watson report.