Add a method to check that MVIDs match
Add a utility method to check that MVIDs match.

Given a path to an assembly and a loaded `Assembly` object,
`AssemblyUtilities.MvidsMatch` returns true if their MVIDs (Module
Version IDs) match, and false otherwise.

This commit contains the code itself and unit tests; a later commit will
wire this into the compilers/VS.