Add a method to check for strong-named assemblies
Add a utility method to check for strong-named assemblies.

`AssemblyUtilities.HasStrongName` uses the metadata reader to check if
the assembly at a given path has a strong name or not.

This commit includes the code itself and a couple of unit tests. A later
commit will wire it into the compilers/VS.