Avoid exceptions for rename conflicts with implicit names
Related to #2511 "Rename doesn't catch conflicts with accessors"

When a rename conflict is introduced by naming something to the name of
a property accessor (e.g. "get_X"), we do not correctly detect that
conflict, and a KeyNotFoundException is thrown (and handled). This fixes
the first chance exception, but does not detect the conflict.