Add global undo to ContainedLanguageCodeSupport.TryRenameElement
Fixes internal TFS bug #1035934 "Undo didn't work in code file after
rename control's ID in properties"

Wrap the contents of ContainedLanguageCodeSupport,.TryRenameElement in a
global undo. This allows the undo of a rename done from the aspx
designer (for example) to correctly revert the C#/VB part of the rename
as well.