Add "Brace Matching" color
Partial fix for internal TFS bug #927616 ""Brace Matching (Highlight)"
Color setting is ignored"

This change adds a new "Brace Matching" color (not to be confused with
the existing "Brace Matching (Highlight)" and "Brace Matching
(Rectangle)" colors) that controls the appearance of brace matching. It
is user customizable and the Background color controls the fill
("Highlight") and the Foreground color controls the outline
("Rectangle").

The default background color was chosen to match Visual Studio 2013's
default "Brace Matching (Highlight)" color.

There will also be a related closed-source change for the themed colors,
and that change will complete the bug.