Introduce new entry named 'Default' in solution explorer context menu
Fixes #1512

The solution explorer context menu that allows users to change severity of analyzer-reported diagnostics currently displays the following entries -
    Error
    Warning
    Info
    Hidden
    None

When the context menu is displayed, a check mark is displayed next to the entry representing the severity of diagnostic as configured in the project's ruleset file. This works fine in cases where a severity is explicitly specified for the diagnostic in the ruleset file.

However, in cases where the severity is not explicitly specified in the ruleset file, the diagnostic is assigned its 'default' severity and we don't display a check mark against any entry in the above menu. As noted in #1512, this is confusing because the 'default' severity is also one of the severities listed above.

This change introduces an additional entry named 'Default' in the above menu that will be checked in cases where the dagnostic does not have an explicitly specified severity in the ruleset file. If the user clicks this entry to change the severity of some diagnostic to 'Default', the entry corresponding to this diagnostic is removed from the ruleset file.