"New" preselection when followed by assignment
C# object creation preselection gets very confused in a situation like:

```
Task x = new $$
string x = "foo"
```
because the new keyword is actually a part of the second assignment
expression. It therefore selects string instead of Task. To avoid this,
the type inferrer now detects if new is on the left of a parenting
assignment expression and on the right of a grandparent assignment
expression.

Object creation preselection now uses the position-based mode of the
TypeInferrer, which needed to be tweaked to properly walk up from
ObjectCreationExpressions.