Add ProvideLanguageCodeExpansionAttribute to packages
Partial fix for internal TFS bug #835188 "The language item of the Code
Snippets Manager was changed"

Add a ProvideLanguageCodeExpansionAttribute to the CSharpPackage and
VisualBasicPackage to specify each languages name as it appears in the
Code Snippets Manager, as well as to specify the SearchPaths and
ForceCreateDirs that were previously specified directly in the setup
authoring.

This change also removes the "Refactoring" snippets from the list of C#
SearchPaths and ForceCreateDirs.

There will be a corresponding closed-source change to remove the C# and
VB snippets-related setup authoring, which will complete the fix for bug
#835188.