Reject builds with analyzer inconsistencies
This commit updates vbcscompiler.exe to reject builds where the list of
analyzers is incomplete (i.e., some of the analyzers depend on
assemblies that are not present), we're unable to load the analyzers, or
the assemblies on disk do not actually match the ones that are loaded.

If one of these checks fails the build request is rejected, and the
clients have been updated to fall back to running the build in their own
process (that is, csc.exe or vbc.exe).

It is possible in some cases that a build that was rejected by
vbcscompiler.exe will succeed in csc.exe/vbc.exe. Generally this will
occur when an analyzer dependency is not passed via the `/analyzer`
switch but happens to be in a place where the CLR will find and load it
automatically. We want to ensure consistency which is why these builds
are not allowed in vbcscompiler.exe; otherwise we risk a build sometimes
succeeding and sometimes failing depending on which assemblies happen to
have been loaded by other, unrelated builds.