Remove `CompilerServerFileWatcher` from vbcscompiler.exe
Now that we have checks in place to ensure that the specified set of
analyzer assemblies are the ones that will actually be used for a given
build, we can remove the `CompilerServerFileWatcher` type entirely. This
type watched for changes to analyzer assemblies on disk, stopped
accepting new build requests when it saw any, and then shut down
vbcscompiler.exe once all the pending builds were complete. The goal was
to avoid using an older-but-already-loaded analyzer assembly when there
was a newer one on disk. The new consistency checks are a more reliable
means of achieving the same ends.

Fixes #2324.
Fixes #2377.
Fixes #2495.